backend/photo-gallery/requirements.txt
boto3>=1.26.0
psycopg2-binary>=2.9.0
PyJWT>=2.8.0